requirements.txt
aiohttp
PyYaml